# docs/swagger.yaml
definitions:
  controller.ResponseHTTP:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  models.Login:
    properties:
      appid:
        type: string
      appsecret:
        type: string
      code:
        type: string
      loginType:
        enum:
        - "1"
        - "2"
        - "3"
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - appid
    - appsecret
    - code
    - loginType
    - password
    - username
    type: object
  models.LoginMobile:
    properties:
      captcha:
        type: string
      mobile:
        type: string
    required:
    - captcha
    - mobile
    type: object
  models.User:
    properties:
      avatarUrl:
        type: string
      code:
        type: string
      created_at:
        type: string
      gender:
        type: string
      id:
        type: integer
      index:
        type: string
      mobile:
        type: string
      nickName:
        type: string
      openid:
        type: string
      password:
        type: string
      region:
        type: string
      unionid:
        type: string
      updated_at:
        type: string
      username:
        description: query tag是query参数别名，json xml，form适合post
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/captcha:
    post:
      consumes:
      - application/json
      description: 获取验证码
      parameters:
      - description: mobile
        in: body
        name: mobile
        required: true
        schema:
          $ref: '#/definitions/models.LoginMobile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取验证码
      tags:
      - 获取验证码
  /v1/login:
    post:
      consumes:
      - application/json
      description: 登录接口
      parameters:
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 登录接口
      tags:
      - 登录
  /v1/region:
    get:
      consumes:
      - application/json
      description: 获取省市区
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取省市区
      tags:
      - 获取省市区
  /v1/upload:
    post:
      consumes:
      - application/json
      description: 文件上传
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 文件上传
      tags:
      - 文件上传
  /v1/user:
    get:
      consumes:
      - application/json
      description: 获取User详情
      parameters:
      - in: query
        name: avatarUrl
        type: string
      - in: query
        name: code
        type: string
      - in: query
        name: created_at
        type: string
      - in: query
        name: gender
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: index
        type: string
      - in: query
        name: mobile
        type: string
      - in: query
        name: nickName
        type: string
      - in: query
        name: openid
        type: string
      - in: query
        name: password
        type: string
      - in: query
        name: region
        type: string
      - in: query
        name: unionid
        type: string
      - in: query
        name: updated_at
        type: string
      - description: query tag是query参数别名，json xml，form适合post
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取User详情
      tags:
      - User
    post:
      consumes:
      - application/json
      description: 添加user
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 添加user
      tags:
      - User
  /v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 删除user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 编辑用户
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 编辑用户
      tags:
      - User
  /v1/users:
    get:
      consumes:
      - application/json
      description: 获取User列表
      parameters:
      - in: query
        name: avatarUrl
        type: string
      - in: query
        name: code
        type: string
      - in: query
        name: created_at
        type: string
      - in: query
        name: gender
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: index
        type: string
      - in: query
        name: mobile
        type: string
      - in: query
        name: nickName
        type: string
      - in: query
        name: openid
        type: string
      - in: query
        name: password
        type: string
      - in: query
        name: region
        type: string
      - in: query
        name: unionid
        type: string
      - in: query
        name: updated_at
        type: string
      - description: query tag是query参数别名，json xml，form适合post
        in: query
        name: username
        type: string
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取User列表
      tags:
      - User
swagger: "2.0"